tool: make some utility test functions reusable
Make them reusable by moving them to the common module.
